requirements.txt
black==23.10.1
flake8==6.1.0
infi.systray==0.1.12
isort==5.12.0
keyboard==0.13.5
pyautogui==0.9.54
setuptools==75.1.0
